smart-task-backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.1
python-decouple==3.8
